requirements.txt
discord.py
pynput
pyyaml
python-dotenv
termcolor
notify2
